8/shop/backend/requirements.txt
Flask
Flask-CORS
Werkzeug
PyJWT
Authlib
requests